requirements.txt
aiohttp==3.9.1
python-telegram-bot==20.7
google-generativeai>=0.8.3
openai==1.6.1
python-dotenv==1.0.0
groq==0.4.2